refactor: Consolidate output packages under pkg/ui
Move pkg/display, pkg/lipbalm, and pkg/output to pkg/ui to centralize
all UI-related functionality. This is Phase 1 of the output rendering
refactor.

No functional changes, just reorganization:
- pkg/display → pkg/ui/display
- pkg/lipbalm → pkg/ui/lipbalm
- pkg/output → pkg/ui/output

All imports updated and tests passing.
