+++ pkg/ui/display/integration_test.go
