+++ pkg/ui/display/simple.go
