+++ pkg/ui/display/simple_test.go
