+++ pkg/ui/output/renderer.go
@@ -8,9 +8,9 @@ import (
 	"os"
 	"text/template"
 
-	"github.com/arthur-debert/dodot/pkg/lipbalm"
 	"github.com/arthur-debert/dodot/pkg/logging"
-	"github.com/arthur-debert/dodot/pkg/output/styles"
+	"github.com/arthur-debert/dodot/pkg/ui/lipbalm"
+	"github.com/arthur-debert/dodot/pkg/ui/output/styles"
 	"github.com/arthur-debert/dodot/pkg/types"
 	"github.com/charmbracelet/lipgloss"
 )
