+++ pkg/ui/output/renderer_test.go
