+++ pkg/ui/output/styles/styles.yaml
