+++ pkg/ui/output/styles/styles_test.go
