+++ pkg/ui/output/templates/pack.tmpl
