+++ pkg/ui/output/templates/result.tmpl
