+++ pkg/ui/output/templates/status-badge.tmpl
