+++ pkg/ui/output/templates/status-inline.tmpl
